0.7.0

+++ jarvis/archive.py
@@ -1,9 +1,10 @@
 import os, pickle, random, time
+from pathlib import Path
 import numpy as np
 from typing import Any, Optional
 
 from .config import Config
-from .utils import progress_str, time_str
+from .utils import tqdm
 from .alias import Array
 
 
@@ -26,16 +27,16 @@ class Archive:
     def __init__(self,
         store_dir: str,
         key_len: int = 8,
-        path_len: int = 2,
+        path_len: int = 4,
         max_try: int = 30,
         pause: float = 0.5,
+        use_buffer: bool = False,
     ):
         r"""
         Args
         ----
         store_dir:
-            The directory for storing data. If `store_dir` is ``None``, an
-            internal dictionary `__store__` is used.
+            The directory for storing data.
         key_len:
             The length of keys.
         path_len:
@@ -45,12 +46,18 @@ class Archive:
             The maximum number of trying to read/write external files.
         pause:
             The time (seconds) between two consecutive read/write attempts.
+        use_buffer:
+            Whether to use a buffer in memory. If ``True``, some methods will be
+            attempted on an internal dict first before loading external files.
+            Overwriting will be disabled since many Archive objects may be
+            running at the same time.
 
         """
         self.store_dir, self.key_len = store_dir, key_len
         assert key_len>=path_len, "File name length should be no greater than key length."
         os.makedirs(self.store_dir, exist_ok=True)
         self.path_len, self.max_try, self.pause = path_len, max_try, pause
+        self.buffer = {} if use_buffer else None
 
     def __repr__(self) -> str:
         return f"Archive object stored in {self.store_dir}."
@@ -58,15 +65,21 @@ class Archive:
     def __setitem__(self, key: str, val: Any):
         store_path = self._store_path(key)
         records = self._safe_read(store_path) if os.path.exists(store_path) else {}
+        if self.buffer is not None:
+            if key in self.buffer:
+                raise RuntimeError(f"Trying to overwrite an existing key {key} in buffer mode.")
+            else:
+                self.buffer[key] = val
         records[key] = val
         self._safe_write(records, store_path)
 
     def __getitem__(self, key: str) -> Any:
         store_path = self._store_path(key)
-        if not os.path.exists(store_path):
-            raise KeyError(key)
-        records = self._safe_read(store_path)
-        return records[key]
+        if self.buffer is None or key not in self.buffer:
+            if not os.path.exists(store_path):
+                raise KeyError(key)
+            records = self._safe_read(store_path)
+        return records[key] if self.buffer is None else self.buffer[key]
 
     def __contains__(self, key: str) -> bool:
         try:
@@ -91,25 +104,36 @@ class Archive:
                 return False
         return True
 
+    @staticmethod
+    def _file_name(key: str, path_len: int) -> str:
+        return '/'.join(key[:path_len])+'.axv'
+
     def _store_path(self, key: str) -> str:
         r"""Returns the path of external file associated with a key."""
         if not self._is_valid_key(key):
             raise KeyError(key)
-        return f'{self.store_dir}/{key[:self.path_len]}.axv'
+        return f'{self.store_dir}/{self._file_name(key, self.path_len)}'
 
-    def _file_names(self) -> list[str]:
-        r"""Returns all valid file names in the directory."""
-        file_names = [
-            f for f in os.listdir(self.store_dir)
-            if f.endswith('.axv') and len(f)==(self.path_len+4)
-        ]
-        random.shuffle(file_names)
-        return file_names
+    @staticmethod
+    def _file_names(folder_name: str, depth: int) -> str:
+        r"""A generator for names of existing records file."""
+        if depth==1:
+            file_names = os.listdir(folder_name)
+            random.shuffle(file_names)
+            for file_name in file_names:
+                if len(file_name)==5 and file_name[0] in Archive._alphabet and file_name.endswith('.axv'):
+                    yield file_name
+        else:
+            subfolder_names = [f for f in os.listdir(folder_name) if f in Archive._alphabet]
+            random.shuffle(subfolder_names)
+            for subfolder_name in subfolder_names:
+                for file_name in Archive._file_names(f'{folder_name}/{subfolder_name}', depth-1):
+                    yield f'{subfolder_name}/{file_name}'
 
     def _store_paths(self) -> list[str]:
         r"""Returns all valid external files in the directory."""
-        store_paths = [f'{self.store_dir}/{f}' for f in self._file_names()]
-        return store_paths
+        for file_name in self._file_names(self.store_dir, self.path_len):
+            yield f'{self.store_dir}/{file_name}'
 
     def _sleep(self):
         r"""Waits for a random period of time."""
@@ -118,8 +142,8 @@ class Archive:
     def _safe_read(self, store_path: str) -> dict:
         r"""Safely reads a file.
 
-        This is designed for cluster use. A MaxTryIOError is raised when too
-        many attempts have failed.
+        This is designed for cluster use, if too many tries have failed, try to
+        delete the file.
 
         """
         count = 0
@@ -135,7 +159,17 @@ class Archive:
             else:
                 break
         if count==self.max_try:
-            raise MaxTryIOError(store_path, count)
+            if os.path.exists(store_path):
+                os.remove(store_path)
+            records = {}
+        if self.buffer is not None:
+            parts = store_path.split('/')[-self.path_len:]
+            parts[-1] = parts[-1][0]
+            head = ''.join(parts)
+            for key in self.buffer:
+                if key.startswith(head) and key not in records:
+                    self.buffer.pop(key)
+            self.buffer.update(records)
         return records
 
     def _safe_write(self, records: dict, store_path: str):
@@ -143,6 +177,7 @@ class Archive:
         count = 0
         while count<self.max_try:
             try:
+                os.makedirs(Path(store_path).parent, exist_ok=True)
                 with open(store_path, 'wb') as f:
                     pickle.dump(records, f)
             except KeyboardInterrupt as e:
@@ -171,6 +206,16 @@ class Archive:
         r"""Removes all data."""
         for store_path in self._store_paths():
             os.remove(store_path)
+
+    def _prune(self) -> None:
+        r"""Removes corrupted files."""
+        max_try, pause = self.max_try, self.pause
+        self.max_try, self.pause = 1, 0.
+        if self.buffer is not None:
+            self.buffer = {}
+        for store_path in self._store_paths():
+            self._safe_read(store_path) # corrupted files are removed in `_safe_read`
+        self.max_try, self.pause = max_try, pause
 
     def keys(self) -> str:
         r"""A generator for keys."""
@@ -201,6 +246,8 @@ class Archive:
 
     def pop(self, key: str) -> Any:
         r"""Removes a specified key and return its value."""
+        if self.buffer is not None:
+            raise RuntimeError("Attempting to pop values in buffer mode.")
         store_path = self._store_path(key)
         if not os.path.exists(store_path):
             return None
@@ -208,45 +255,37 @@ class Archive:
         val = records.pop(key, None)
         if records:
             self._safe_write(records, store_path)
-        else:
+        elif os.path.exists(store_path):
             os.remove(store_path) # remove empty external file
         return val
 
-    def prune(self) -> list[str]:
-        r"""Removes corrupted files.
+    def delete(self, keys: list[str]) -> None:
+        r"""Removes keys in batch processing."""
+        if self.buffer is not None:
+            raise RuntimeError("Attempting to delete keys in buffer mode.")
+        to_remove = {}
+        for key in keys:
+            store_path = self._store_path(key)
+            if store_path in to_remove:
+                to_remove[store_path].append(key)
+            else:
+                to_remove[store_path] = [key]
+        for store_path in tqdm(to_remove, unit='file', leave=False):
+            if not os.path.exists(store_path):
+                continue
+            records = self._safe_read(store_path)
+            modified = False
+            for key in to_remove[store_path]:
+                if key in records:
+                    records.pop(key)
+                    modified = True
+            if modified:
+                if records:
+                    self._safe_write(records, store_path)
+                elif os.path.exists(store_path):
+                    os.remove(store_path)
 
-        Returns
-        -------
-        keys:
-            The keys of remaining valid records.
-
-        """
-        valid_keys, removed = [], []
-        store_paths = self._store_paths()
-        verbose, tic = None, time.time()
-        for i, store_path in enumerate(store_paths, 1):
-            try:
-                records = self._safe_read(store_path)
-                valid_keys += list(records.keys())
-            except:
-                print("{} corrupted, will be removed".format(store_path))
-                os.remove(store_path)
-                removed.append(store_path[(-4-self.path_len):-4])
-            if i%(-(-len(store_paths)//10))==0 or i==len(store_paths):
-                toc = time.time()
-                if verbose is None:
-                    # display progress if estimated time is longer than 20 mins
-                    verbose = (toc-tic)/i*len(store_paths)>1200
-                if verbose:
-                    print("{} ({})".format(
-                        progress_str(i, len(store_paths)),
-                        time_str(toc-tic, progress=i/len(store_paths)),
-                    ))
-        if removed:
-            print(f"{len(removed)} corrupted files removed.")
-        return valid_keys
-
-    def copy_to(self,
+    def migrate(self,
         dst_dir: str,
         keys: Optional[set[str]] = None,
         overwrite: bool = False,
@@ -264,42 +303,65 @@ class Archive:
 
         """
         os.makedirs(dst_dir, exist_ok=True)
-        # check key consistency of existing axv files
-        for file_name in os.listdir(dst_dir):
-            if file_name.endswith('.axv'):
-                assert len(file_name)==(self.path_len+4), (
-                    f"Data in the destination '{dst_dir}' have inconsistent file name length "
-                    f"{self.path_len}."
-                )
-                _records = self._safe_read(f'{dst_dir}/{file_name}')
-                _key = next(iter(_records.keys()))
-                assert len(_key)==self.key_len, (
-                    f"Data in the destination '{dst_dir}' have inconsistent key length "
-                    f"{self.key_len}."
-                )
-        # merge records
-        for file_name in self._file_names():
-            try:
-                src_path = f'{self.store_dir}/{file_name}'
-                dst_path = f'{dst_dir}/{file_name}'
-
+        # check key consistency
+        path_len = []
+        for depth in range(1, self.key_len+1):
+            file_name =  next(iter(self._file_names(dst_dir, depth)), None)
+            if file_name is not None:
+                path_len.append(depth)
+        if len(path_len)>1:
+            raise RuntimeError(f"Multiple hierarchies detected in {dst_dir}")
+        elif len(path_len)==1:
+            dst_path_len = path_len[0]
+            file_name =  next(iter(self._file_names(dst_dir, dst_path_len)), None)
+            records = self._safe_read(f'{dst_dir}/{file_name}')
+            key = next(iter(records.keys()))
+            assert self._is_valid_key(key), f"Invalid key detected in {dst_dir} ({key})"
+        else:
+            dst_path_len = self.path_len
+        # prepare generator of source file paths
+        if keys is None:
+            src_paths = list(self._store_paths())
+        else:
+            src_paths = set()
+            for key in keys:
+                src_paths.add(self._store_path(key))
+            src_paths = list(src_paths)
+        # copy records to destination directory
+        if dst_path_len>=self.path_len:
+            random.shuffle(src_paths)
+            for src_path in tqdm(src_paths, unit='file', leave=False):
                 src_records = self._safe_read(src_path)
-                src_records = {k: v for k, v in src_records.items() if keys is None or k in keys}
-                dst_records = self._safe_read(dst_path) if os.path.exists(dst_path) else {}
-                if not overwrite:
-                    for key in src_records:
-                        assert key not in dst_records, (
-                            f"Existing key '{key}' detected in the destination '{dst_dir}', cloning "
-                            "operation aborted."
-                        )
-                dst_records.update(src_records)
-                if dst_records:
-                    self._safe_write(dst_records, dst_path)
-            except MaxTryIOError as e:
-                print(str(e))
+                src_keys = [k for k in src_records if keys is None or k in keys]
+                key_dicts = {}
+                for key in src_keys:
+                    _key = key[:dst_path_len]
+                    if _key in key_dicts:
+                        key_dicts[_key].append(key)
+                    else:
+                        key_dicts[_key] = [key]
+                for dst_keys in key_dicts.values():
+                    dst_path = f'{dst_dir}/{self._file_name(dst_keys[0], dst_path_len)}'
+                    if os.path.exists(dst_path):
+                        dst_records = self._safe_read(dst_path)
+                        modified = False
+                    else:
+                        dst_records = {}
+                        modified = True
+                    for key in dst_keys:
+                        if key not in dst_records or overwrite:
+                            dst_records[key] = src_records[key]
+                            modified = True
+                    if modified:
+                        self._safe_write(dst_records, dst_path)
+        else:
+            raise NotImplementedError
 
 
 class HashableRecordArchive(Archive):
+
+    def __init__(self, *args, path_len: int = 2, use_buffer: bool = True, **kwargs):
+        super().__init__(*args, path_len=path_len, use_buffer=use_buffer, **kwargs)
 
     @classmethod
     def _to_hashable(cls, n_val):
@@ -342,7 +404,8 @@ class HashableRecordArchive(Archive):
         return cls._to_hashable(n_val_0)==cls._to_hashable(n_val_1)
 
     def __setitem__(self, key: str, n_val: Any):
-        super().__setitem__(key, self._to_hashable(n_val))
+        h_val = self._to_hashable(n_val)
+        super().__setitem__(key, h_val)
 
     def __getitem__(self, key: str) -> Any:
         return self._to_native(super().__getitem__(key))
@@ -358,6 +421,10 @@ class HashableRecordArchive(Archive):
     def get_key(self, n_val) -> Optional[str]:
         r"""Returns the key of a record."""
         h_val = self._to_hashable(n_val)
+        if self.buffer is not None:
+            for key, val in self.buffer.items():
+                if val==h_val:
+                    return key
         for key, val in super().items():
             if val==h_val:
                 return key
@@ -393,12 +460,8 @@ class HashableRecordArchive(Archive):
 
 class ConfigArchive(HashableRecordArchive):
 
-    def __init__(self,
-        store_dir: str,
-        max_try: int = 300,
-        **kwargs,
-    ):
-        super().__init__(store_dir, max_try=max_try, **kwargs)
+    def __init__(self, *args, max_try: int = 300, **kwargs):
+        super().__init__(*args, max_try=max_try, **kwargs)
 
     def _to_native(self, h_val):
         n_val = super()._to_native(h_val)
@@ -436,3 +499,28 @@ class ConfigArchive(HashableRecordArchive):
             for f_dict in f_dicts
         ]
         return shared, diffs
+
+    def filter(self, cond: dict) -> str:
+        r"""Generator of matching record key.
+
+        Args
+        ----
+        cond:
+            Filter conditions, specifying parts of config values.
+
+        """
+        f_cond = Config(cond).flatten()
+        for key, config in self.items():
+            f_config = config.flatten()
+            matched = True
+            for k, v in f_cond.items():
+                if not (
+                    k in f_config and (
+                        (callable(v) and v(f_config[k])==True) or
+                        f_config[k]==v
+                    )
+                ):
+                    matched = False
+                    break
+            if matched:
+                yield key
